gemini-computer-use-android/requirements.txt
google-genai==1.52.0
python-dotenv==1.2.1
